# openapi.yaml
openapi: 3.0.3
info:
  title: Midori API
  version: 0.1.0
  description: >
    OpenAPI specification สำหรับระบบ Midori - AI-powered website generation platform.
    รองรับการจัดการ authentication, projects, templates, conversations, และ deployment.
  contact:
    name: Midori Team
    url: https://midori.local
  license:
    name: Proprietary
servers:
  - url: http://localhost:3000
    description: Local development server
  - url: https://api.midori.production
    description: Production server (ถ้ามี)
...
components:
  securitySchemes:
    sessionCookie:
      type: apiKey
      in: cookie
      name: midori-session
      description: Session cookie สำหรับ authentication
  schemas:
    Error:
      type: object
      properties:
        success:
          type: boolean
          example: false
        error:
          type: string
          example: เกิดข้อผิดพลาด
        details:
          type: object
          additionalProperties: true
    User:
      type: object
      properties:
        id:
          type: string
          format: uuid
        email:
          type: string
          format: email
        displayName:
          type: string
          nullable: true
        avatarUrl:
          type: string
          format: uri
          nullable: true
        createdAt:
          type: string
          format: date-time
        lastLoginAt:
          type: string
          format: date-time
          nullable: true
...
paths:
...
  /api/auth/login:
    post:
      tags: [Authentication]
      summary: เข้าสู่ระบบ
      description: Login ด้วย email และ password
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [email, password]
              properties:
                email:
                  type: string
                  format: email
                  example: user@example.com
                password:
                  type: string
                  format: password
                  example: password123
      responses:
        '200':
          description: เข้าสู่ระบบสำเร็จ
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  message:
                    type: string
                    example: เข้าสู่ระบบสำเร็จ
                  user:
                    $ref: '#/components/schemas/User'
        '400':
          description: ข้อมูลไม่ถูกต้อง
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '401':
          description: Email หรือ password ไม่ถูกต้อง
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /api/auth/register:
    post:
      tags: [Authentication]
      summary: สมัครสมาชิก
      description: สร้างบัญชีผู้ใช้ใหม่
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [email, password, displayName]
              properties:
                email:
                  type: string
                  format: email
                password:
                  type: string
                  format: password
                  minLength: 8
                displayName:
                  type: string
                  minLength: 2
      responses:
        '200':
          description: สมัครสมาชิกสำเร็จ
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  message:
                    type: string
                  user:
                    $ref: '#/components/schemas/User'
        '400':
          description: ข้อมูลไม่ถูกต้องหรือ email ซ้ำ
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /api/auth/logout:
    post:
      tags: [Authentication]
      summary: ออกจากระบบ
      security:
        - sessionCookie: []
      responses:
        '200':
          description: ออกจากระบบสำเร็จ
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  message:
                    type: string
        '401':
          description: ไม่พบ session
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
...
  /api/generate:
    get:
      tags: [Generate]
      summary: สร้างเว็บไซต์จาก keywords
      description: ใช้ AI สร้างเว็บไซต์จาก business keywords และ override system
      parameters:
        - name: keywords
          in: query
          schema:
            type: string
            example: fresh,organic,homemade
          description: Keywords คั่นด้วย comma
      responses:
        '200':
          description: สร้างเว็บไซต์สำเร็จ
          content:
            application/json:
              schema:
                type: object
                properties:
                  files:
                    type: array
                    items:
                      type: object
                      properties:
                        path:
                          type: string
                        content:
                          type: string
                  businessCategory:
                    type: string
                  concreteManifest:
                    type: object
                  appliedOverrides:
                    type: array
                    items:
                      type: string
                  processingTime:
                    type: number
                  validationResults:
                    type: object
        '500':
          description: เกิดข้อผิดพลาด

  # ========== Visual Edit ==========
...
  /api/preview/daytona:
    post:
      tags: [Preview]
      summary: สร้าง Daytona sandbox ใหม่
      description: สร้าง sandbox สำหรับ preview โปรเจกต์
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [projectId, files]
              properties:
                projectId:
                  type: string
                  format: uuid
                files:
                  type: array
                  items:
                    type: object
                    properties:
                      path:
                        type: string
                      content:
                        type: string
      responses:
        '200':
          description: สร้าง sandbox สำเร็จ
          content:
            application/json:
              schema:
                type: object
                properties:
                  sandboxId:
                    type: string
                  url:
                    type: string
                    format: uri
                  state:
                    type: string

    get:
      tags: [Preview]
      summary: ดึงสถานะ sandbox
      description: ตรวจสอบสถานะและ heartbeat ของ sandbox
      parameters:
        - name: sandboxId
          in: query
          required: true
          schema:
            type: string
      responses:
        '200':
          description: สถานะ sandbox
          content:
            application/json:
              schema:
                type: object
                properties:
                  sandboxId:
                    type: string
                  state:
                    type: string
                  url:
                    type: string
        '400':
          description: ไม่มี sandboxId
        '404':
          description: ไม่พบ sandbox

    put:
      tags: [Preview]
      summary: อัปเดตไฟล์ใน sandbox
      parameters:
        - name: sandboxId
          in: query
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                files:
                  type: array
                  items:
                    type: object
                    properties:
                      path:
                        type: string
                      content:
                        type: string
      responses:
        '200':
          description: อัปเดตสำเร็จ
        '400':
          description: ไม่มี sandboxId
        '404':
          description: ไม่พบ sandbox

    delete:
      tags: [Preview]
      summary: ลบ sandbox
      parameters:
        - name: sandboxId
          in: query
          required: true
          schema:
            type: string
      responses:
        '200':
          description: ลบสำเร็จ
        '400':
          description: ไม่มี sandboxId
        '500':
          description: ลบล้มเหลว